+++ src/metadata/ffmpeg_thumbnailer_handler.h
@@ -34,12 +34,23 @@ Gerbera - https://gerbera.io/
 
 class CdsObject;
 class IOHandler;
+enum class ObjectType;
 
 class FfmpegThumbnailerHandler : public MediaMetadataHandler {
 public:
-    explicit FfmpegThumbnailerHandler(const std::shared_ptr<Context>& context, ConfigVal checkOption);
+    explicit FfmpegThumbnailerHandler(
+        const std::shared_ptr<Context>& context,
+        ConfigVal checkOption,
+        ObjectType mediaType);
+
+    bool isSupported(const std::string& contentType,
+        bool isOggTheora,
+        const std::string& mimeType,
+        ObjectType mediaType) override;
     bool fillMetadata(const std::shared_ptr<CdsObject>& obj) override;
-    std::unique_ptr<IOHandler> serveContent(const std::shared_ptr<CdsObject>& obj, const std::shared_ptr<CdsResource>& resource) override;
+    std::unique_ptr<IOHandler> serveContent(
+        const std::shared_ptr<CdsObject>& obj,
+        const std::shared_ptr<CdsResource>& resource) override;
 
 protected:
     // Needed in tests
@@ -51,6 +62,7 @@ private:
     // Add a lock around the usage to avoid crashing randomly.
     mutable std::mutex thumb_mutex;
     fs::path cachePath;
+    ObjectType mediaType;
     std::optional<std::vector<std::byte>> readThumbnailCacheFile(const fs::path& movieFilename) const;
     void writeThumbnailCacheFile(const fs::path& movieFilename, const std::byte* data, std::size_t size) const;
 };