+++ src/metadata/matroska_handler.cc
@@ -118,6 +118,15 @@ MatroskaHandler::MatroskaHandler(const std::shared_ptr<Context>& context)
 {
 }
 
+bool MatroskaHandler::isSupported(
+    const std::string& contentType,
+    bool isOggTheora,
+    const std::string& mimeType,
+    ObjectType mediaType)
+{
+    return contentType == CONTENT_TYPE_MKV;
+}
+
 bool MatroskaHandler::fillMetadata(const std::shared_ptr<CdsObject>& obj)
 {
     auto item = std::dynamic_pointer_cast<CdsItem>(obj);